% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/districts_example.R
\name{districts_example}
\alias{districts_example}
\title{Example district-level enrollment data}
\format{
A data frame with 10 rows and 5 columns
}
\description{
Example district-level enrollment data
}
\examples{
data(districts_example)
head(districts_example)
}
\keyword{internal}
