% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_assessment_via_portal}
\alias{download_assessment_via_portal}
\title{Download assessment via MDE portal}
\usage{
download_assessment_via_portal(end_year, level, dest_file)
}
\arguments{
\item{end_year}{School year end}

\item{level}{Data level}

\item{dest_file}{Destination file path}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Downloads assessment data using the MDEAnalytics portal interface.
This mimics the web interface download functionality.
}
\keyword{internal}
