% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_mde_enrollment}
\alias{download_mde_enrollment}
\title{Download MDE enrollment data file}
\usage{
download_mde_enrollment(end_year, level = "school")
}
\arguments{
\item{end_year}{School year end}

\item{level}{Data level: "school", "district", or "state"}
}
\value{
Data frame with enrollment data
}
\description{
Downloads enrollment data from the MDEAnalytics portal.
MDE provides Excel files for download that can be accessed via direct URLs.
}
\keyword{internal}
