% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr_2024_example.R
\docType{data}
\name{enr_2023_example}
\alias{enr_2023_example}
\title{Example enrollment data (2023)}
\format{
A data frame with enrollment data
}
\source{
Minnesota Department of Education, MARSS enrollment data
}
\description{
Real enrollment data from the Minnesota Department of Education for the
2022-23 school year. Contains district-level enrollment by race/ethnicity
for all 389 Minnesota districts (including 68 charter schools).
}
\keyword{internal}
