% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr_2024_example.R
\name{enr_2024_example}
\alias{enr_2024_example}
\title{Example 2024 enrollment data}
\format{
A data frame with 15 rows and 9 columns
}
\description{
Example 2024 enrollment data
}
\examples{
data(enr_2024_example)
head(enr_2024_example)
}
\keyword{internal}
