% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr_multi_example.R
\docType{data}
\name{enr_multi_example}
\alias{enr_multi_example}
\title{Example multi-year enrollment data}
\format{
A data frame with enrollment data
}
\source{
Minnesota Department of Education, MARSS enrollment data
}
\description{
Real enrollment data from the Minnesota Department of Education.
Currently contains 2022-23 school year data.
}
\keyword{internal}
