% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{get_assessment_cache_path}
\alias{get_assessment_cache_path}
\title{Get assessment cache file path}
\usage{
get_assessment_cache_path(end_year, type, level)
}
\arguments{
\item{end_year}{School year end}

\item{type}{Data type ("tidy" or "wide")}

\item{level}{Data level ("all", "state", "district", "school")}
}
\value{
Full path to cache file
}
\description{
Get assessment cache file path
}
\keyword{internal}
