% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{get_available_assessment_years}
\alias{get_available_assessment_years}
\title{Get available assessment years}
\usage{
get_available_assessment_years()
}
\value{
List with 'years' (numeric vector) and 'note' (character)
}
\description{
Returns information about years for which assessment data is available.
}
\examples{
get_available_assessment_years()
}
