% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_directory_url}
\alias{get_directory_url}
\title{URLs for MDE-ORG directory extracts}
\usage{
get_directory_url(type = "school")
}
\arguments{
\item{type}{Either "school" or "district"}
}
\value{
URL string for CSV extract
}
\description{
Returns the extract URLs for school and district directory data from MDE-ORG.
}
\keyword{internal}
