% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{get_raw_assessment}
\alias{get_raw_assessment}
\title{Download raw assessment data from MDE}
\usage{
get_raw_assessment(end_year, level = "all")
}
\arguments{
\item{end_year}{School year end (e.g., 2024 for 2023-24 school year)}

\item{level}{Data level: "all" (default), "state", "district", or "school"}
}
\value{
List with state, district, and/or school data frames
}
\description{
Downloads assessment data from Minnesota Department of Education's
MDEAnalytics portal.
}
\keyword{internal}
