% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_district_directory}
\alias{get_raw_district_directory}
\title{Download raw district directory data from MDE-ORG}
\usage{
get_raw_district_directory()
}
\value{
Data frame with raw district directory data
}
\description{
Downloads the district (LEA) directory CSV extract from the MDE Organization
Reference Glossary (MDE-ORG).
}
\keyword{internal}
