% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_school_directory}
\alias{get_raw_school_directory}
\title{Download raw school directory data from MDE-ORG}
\usage{
get_raw_school_directory()
}
\value{
Data frame with raw school directory data
}
\description{
Downloads the school directory CSV extract from the MDE Organization
Reference Glossary (MDE-ORG).
}
\keyword{internal}
