% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{read_directory_cache}
\alias{read_directory_cache}
\title{Read directory data from cache}
\usage{
read_directory_cache(end_year, type)
}
\arguments{
\item{end_year}{School year end}

\item{type}{Cache type}
}
\value{
Cached data
}
\description{
Read directory data from cache
}
\keyword{internal}
