% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{standardize_test_name}
\alias{standardize_test_name}
\title{Standardize test names}
\usage{
standardize_test_name(x)
}
\arguments{
\item{x}{Vector of test names}
}
\value{
Standardized test names
}
\description{
Standardize test names
}
\keyword{internal}
