% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/districts_example.R
\docType{data}
\name{districts_example}
\alias{districts_example}
\title{Example districts list}
\format{
A data frame with district names and enrollment counts
}
\source{
Minnesota Department of Education, MARSS enrollment data
}
\description{
Total enrollment by district for the 2022-23 school year from MDE.
}
\keyword{internal}
