% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr_multi_example.R
\name{enr_multi_example}
\alias{enr_multi_example}
\title{Example multi-year enrollment data (2022-2024)}
\format{
A data frame with 45 rows and 12 columns
}
\description{
Example multi-year enrollment data (2022-2024)
}
\examples{
data(enr_multi_example)
head(enr_multi_example)
}
\keyword{internal}
